requirements.txt
Flask==2.3.3
google-cloud-ndb==2.2.1
google-cloud-datastore==2.18.1
Werkzeug==2.3.7
python-dotenv==1.0.0
